fix: image pull via the API should not have timeout or retries
Fixes #13948

There is a bit of moving code around for better readability, but the
tl;dr of the changes is:

* the new `Pull` method doesn't do any timeouts or retries on top of
  whatever `ctx` passed in, it is used in the API
* the new method `PullWithRetriesAndTimeout` matches previous `Pull`
  method semantics - automatically retries and provides global pull
  timeout - this is used internally when pulling etcd/kubelet images, in
  the legacy Upgrade API.

Also for the new pull API a proper logger is used, so that containerd
client logs don't flow to the console.

There should be no functional changes execept for image pull APIs to
have no timeout/retries.
