+++ internal/app/machined/internal/server/v1alpha1/v1alpha1_images.go
@@ -102,7 +102,6 @@ func (s *Server) ImagePull(ctx context.Context, req *machine.ImagePullRequest) (
 		s.Controller.Runtime().State().V1Alpha2().Resources(),
 		client, req.Reference,
 		image.WithSkipIfAlreadyPulled(),
-		image.WithMaxNotFoundRetries(0), // return an error immediately if the image is not found
 	)
 	if err != nil {
 		if errdefs.IsNotFound(err) {
